include permutations, save first_level, response as extra regressor
Major overhaul of MEG inference script:

1) You can now give the number of permutations which should be computed and
stored together with the genuine inference results in a common file. The genuine
results are in permutation number 0.

2) I now also save the results of the first level, i.e., for individual
subjects. This includes betas, their standard errors and some diagnostics such
as the likelihood of the fitted GLM. It now takes longer to compute the results
which must be due to computation of standard errors or the diagnostics.
Subjectively, I would guess that computation time has doubled. I left the
second_level variable in for convenience, but all it contains can be computed
from the first_level results, if desired.

3) I included response as a new regressor, because it is pretty much
uncorrelated with the other used regressors and gives potentially relevant new
information. As entropy and trial_time, response does not vary with dots. Hence,
the betas provide something like a moving average over the 5 100 ms separated
time points within a trial of the amount of correlation of the MEG signal with
the response across trials.

4) Only use single time point for trial regressors: Trial regressors are
entropy, response and trial_time which only vary across trials, but not dots. I
have introduced an option to not average over the 5 dot time points in the
defined sequence, but apply them only to the time point of a single dot which is
chosen with trialregs_dot (set to 0 for standard averaging over the chosen
sequence). For example, for trialregs_dot=5 you only look for an effect of the
trial regressors at the given t0 from 5th dot onset. I introduced this, because
I thought that the averaging actually reduces observable effects for the trial
regressors which all tend to increase towards the end of the considered time
window (900 ms), especially for response. It turns out, however, that averaging
across 500 ms leads to clearer effects even for the trial regressors.

5) Results files are now time stamped for better archival.

Results
-------
I produced two major results which both contain the genuine inference results
plus 5 permutations (with standard options, i.e., with baseline correction,
downsampling to 100Hz, without smoothing, first 5 dots in sequence):
- meg_sequential_201703011624.h5 (averaging trial regressors, trial_regs=0)
- meg_sequential_201703011927.h5 (trial_regs=5)

Statistical significance
........................
I have only 5 permutations, but these clearly show that the inferred means are
highly significant, because shuffling the trials (differently at each time
point) leads to mean betas that stay below 0.02 (except for abs_dot_x for which
permuted mean betas can be as high as 0.03) when selecting the (absolute)
maximum across channels while the major effect reach mean betas above 0.04.

Difference in intercept across permutations
...........................................
I permute only the trials within the different time points. This should not lead
to differences in the inferred intercept, which is the mean across trials,
across permutations. Yet, I see small deviations of the curves. Specifically,
the permuted mean betas have a small bump at 410 ms which is not there in the
genuine results. First I thought that the differences can only result from me
excluding trials after the permutation based on either too fast RT of the
subject or too few simulations for that trial, but this is not true. I have
checked that the permutations and trial selection works as intended, but still I
see subjects which exhibit strong differences in intercept between permutations
although only very few trials are excluded (e.g., subject 15 who only has 1
excluded trial).

It may be that this is an effect carrying over from the other regressors, i.e.,
it may be that in the genuine results the other regressors actually explain a
little bit of the average captured by the intercept and that this goes away with
permuting, or the other way around depending on whether the permuted intercept
is higher or lower.

A quick look through some subjects suggests that the amount of differences in
intercept may be related to the size of the fitted noisestd, but this is
inconclusive and I need to check this more thoroughly.

Single subjects
...............
differ widely in their results, but the sequential approach still allows to see
reliable single-subject effects.

Subject 4 sticks out as exhibiting remarkably clear effects which are very
similar to the grand average results. It also shows a gradual increase of the
intercept towards the end of the trial. Other subjects show this, too, and I
wonder why this cannot be seen in the grand average, but there are also subjects
for which the grand average decreases.

It appears that the intercept of single subjects is dominated by one of the
sequential peaks - either 20 ms or 60 ms, but this also depends on whether you
look at the maximum positive or negative effects which can differ considerably.
In some subjects, e.g., 16, it appears that the positive maxima show the 60 ms
peaks while the negative maxima show the 20 ms peaks. In other subjects, e.g.,
4, positive and negative peaks align. I intended to look into this more
rigorously, but didn't get very far so far.