new mode: absmean, supposedly closer to overall power of reconstructed sources
I added a new mode for showing the activation over channels for a fixed time
point: 'absmean' which takes the absolute value of the measure and then averages
it across all channels. I added this, because it appeared to me that the peaks
in the absmax plots don't necessarily appear at times where there also is the
greatest amount of overall activation in source space, but this was before I
went back to using the channel-specific stds and dSPM (see one of the next
commits about source reconstruction). All in all, I think that there is a
reasonably good correspondence between the peaks in sensor space and source
space activations. So using absmean may not be necessary and absmax may be
easier interpretable (the mean is relatively low, too).