function for extracting time courses for labels in source space
Given a previously computed source estimate and a brain parcellation this code
will extract the time courses of average activation in the labelled regions
defined in the parcellation.

I chose the new HCPMMP1 parcellation from Glasser2016 which is available for
fsaverage through MNE (note: license requires citation - should be natural to
comply).

MNE provides different ways of averaging. I chose the standard: mean_flip which
supposedly flips the sign of the source current based on the dipole orientation.
I guessed that MNE needs the orientation information of the source space and
therefore built the source space from the forward model exactly as I did during
source reconstruction. It appears a bit of a waste to me to essentially throw
away the dipole orientations again after taking care to have them during source
reconstruction, but it's definitely necessary for sensibly averaging within
arbitrary label regions.

I'm also a bit wary about throwing away information about the variability of the
signal within the label regions. MNE only returns the means, but not the stds.
At least I save the number of signal points that contributed to the average for
each region.