conducting second-level inference for extracted label time courses
Because I use consistent variable names, necessary changes in code were minimal.
I only had to take care of different lengths of label names in the index of
first_level_src. Previously this was no issue, because the corresponding index
level only contained vertex numbers, not label strings.

I tested on dot_x. Hence, this is not in the list of r_names committed here.

It doesn't make much sense to allow sign flips across subjects for the
regressors which have nothing to do with the arbitrary assignment of left-right
in the decision, but I run the analysis for all regressors anyway to have
consistent results, because it doesn't hurt and because it only takes ~1 hour
per regressor for the extracted label time courses now.