show areas as data, identify consistent sequences, exploration
I have found a way to show areas (also called labels in the code) as data on the
brain without smoothing. The simple trick is to get the vertices of the
visualised surface belonging to a particular area and just setting those to a
common value. This allowed me to explore the results for the areas and I
realised that there's a lot happening which I cannot interpret reliably. For
example, there appear to be spikes in some areas which reach quite high
mu_p_large, but only for one time point and then drop back to a low value. I
doubt that these effects are genuine. So I devised a criterion that is supposed
to identify reliably activated areas as described next.

identifying consistent sequences
--------------------------------
It's clear that I'm only interested in the activations that are above some
threshold in mu_p_large. As a kind of Baysesian reverse p-value a threshold of
0.95 appears to be reasonable, but this has to be interpreted in relation to my
choice of 0.5 in the definition of mu_p_large. I could increase all mu_p_large
values simply by reducing 0.5 to 0.3.

So instead I though of another criterion: I pool all mu_p_large values of all
regressors, labels and time points to one distribution and select the threshold
as the value that corresponds to the 0.95 quantile of that distribution. So I
take the data distribution as my null distribution. The idea being that I
disregard all values which I could easily draw randomly from that distribution.

Interestingly, the threshold value identified by this criterion is with ~0.75
relatively low, but, looking at single label time courses, it appears to be
justified, because activations tend to rise from some low value below 0.5 to
above 0.9 for regions which reach these high values. This low threshold of 0.75,
however, also means that quite a few seemingly unreliable fluctuations cross the
threshold. Also, you should remember that I do a vast number of hypothesis tests
when identifying activations (360 areas * 50 time points = 18000 tests). So
naively I would expect 900 false positives with this procedure. This number is
hard to guess in reality, though, because I estimated the 0.95 threshold from
the data distribution which contains the true effects. Depending on the unknown
number of true effects and how much the true effect distribution overlaps with
the null distribution, the number of false positives could be considerably lower
and in some situations even higher than the naive 900. In general, however,
under the assumption that the true effects tend to have higher values than the
null effects, the expected number of false positives should be lower than 900
under my data distribution criterion.

Note that this expected number of errors relates to the per comparison error
rate (PCER) which is defined as the expected number of null hypotheses which are
declared significant divided by the total number of tests. The false discovery
rate (FDR), on the other hand, is the expected ratio of false positives within
the set of accepted hypotheses, i.e., significant tests. Finally, the
family-wise error rate (FWER) is defined as the probability of having at least
one false positive among all tests. See Benjamini1995.

To reduce the expected error rate and because I presume that single-time point
effects are spurious, I extended the criterion to only include effects that
occurred in a contiguous sequence. I set the desired sequence length arbitrarily
to 3 to tradeoff my sensitivity to detect short activations (now need to be at
least 30 ms long) with my specificity to exclude spurious effects. Consequently,
the PCER drops from 0.05 to 0.05^3=0.000125 meaning that I only expect ~2 false
positives per 18000 tests. Basis of this argument is that random effects in
sequential time points are independent, i.e., the noise is not correlated. I
have some doubt that this is true, but my hope is that all the previous
processing steps have reduced these inter-temporal noise correlations
sufficiently such that it is no issue at this stage anymore. It's certainly the
case that all my previous analyses assume independent time points.

Some results
------------
Only showing activations that fulfill the above criterion appears to give a more
interpretable picture in the sense that it simply looks more natural, when
areas steadily increase their activation, stay a bit activated and decrease
their activation again. There is still a lot going on, though, and it's hard to
keep track of it.

In the hope that it would be simpler to interpret the activation of whole
regions I also now provide the option to show the activation in the 22 sections
defined in Glasser2016 (e.g. based on the maximum activation in the
corresponding brain region). When plotting the regions on the brain, this is not
really the case. Perhaps I should rather plot the time courses of the regions
instead.