Correcting associations between Glasser's areas and sections
Checking some data I noticed that the bold face section numbers in the table of
Glasser2016 are not always consistent with the association between areas and
sections in the actual text. So I extracted the areas discussed under each
section in the text and corrected the associations in Glasser2016_areas.csv. The
changes can be seen in the diff of the file.

The inconsistencies that triggered my investigation was the activation of early
auditory cortex (section 10) around 420 ms after dot onset. These were driven by
areas 52 and PFcm which are now associated with different sections.