plotting region time courses instead of brain
Even with the aggregation into regions, there are still 12 (left) and 14 (right)
regions activated for the dot_x regressor which makes results still unclear. To
counter that I plot only maximally 3 time courses in one panel. This at least
allows to identify individual time courses and I can start to interpret them.

In the time courses for dot_x I recognise the activations at 120 and 170 ms, but
the 120 ms peaks appear to be comparably weak. I also see the sustained
activations from 300 ms onwards. Additionally, however, there are also effects
in between 170 and 300 ms which I would like to interpret as the interesting
computations that actually lead to the updating of the accumulated evidence
supposedly reflected after 300 ms. I will need to check the sequence of peaks
across regions in that time slot to make this more concrete.

I also briefly checked the results for the accev regressor. These slightly
differ for the left and right hemisphere (as for dot_x). However, both
hemispheres have in common that anterior cingulate and medial prefrontal cortex
as well as posterior cingulate cortex show sustained activation through almost
the whole time period. In the right hemisphere this is also true for the dorsal
stream visual cortex. The other regions show intermittent activation and I
haven't identified a particular pattern in there, yet.

What does an intermittent representation of accumulated evidence in an area
mean? Given that accumulated evidence is correlated with dot_x what are the
patterns that you should expect to see in accumulated evidence based on the
patterns in dot_x and when would you not expect to see dot_x patterns in accev?
In any case, if dot_x patterns are to be found in accev, then they should occur
100 ms earlier assuming that accev mostly reflects the dot_x patterns of the
previous dot.