adding choice-flipped dot_x and accev regressors
This is motivated by an analysis in Cheadle2014 in which they separate trials
by the made choices and show that the decision-relevant information in the
stimulus correlates positively for the one choice and negatively for the other
with BOLD in some selected brain areas. This is the same as flipping the sign of
the stimulus information in the regressor for trials in which one of the choices
was made. The resulting regressor does not correlate with the original
information regressor anymore. The whole thing doesn't make much sense, because
there shouldn't be an influence of future choice on the processing of stimulus
information that supposedly leads to that choice. So I expected no significant
activation for the new dot_x_cflip regressor, but wasn't so sure about
accev_cflip, because accev eventually represents choice and who knows.

So I implemented accev_cflip and dot_x_cflip and added them to the GLM. Results
are stored in

    meg_sequential_201705311314.h5

where I chose to not baseline correct and used trialregs_dot = 0.

Regressor correlations
----------------------
accev_cflip of the first dot is equal to response, because accev for the first
dot only contains an offset that is equal across trials. For subsequent dots
accev_cflip becomes quite strongly correlated with entropy and I wonder what
that means! Also dot_x_cflip has some correlation with entropy for later dots.
This is strange as entropy itself has no correlation with the response. Clearly,
dot_x of the previous dot is (anti-)correlated with accev of the current dot.
This is also true for the flipped versions, but compiling all of the single dots
into one regressor solves that.

Results
-------
As expected, dot_x_cflip has no significant effects, but only initially. From
about 300 ms there is a rise in absmean across channels. Perhaps this is
something driven by only the 5th dot, or so, when the choice becomes reflected
in the activity. I should check a single dot analysis.

accev_cflip, on the other hand, shows similar, but lower effects as accev. What
is weird there is that the permutations don't break all of these effects, i.e.,
the permuted absmean trajectories strangely show the same peaks, a bit smaller,
as the unpermuted results. I also don't know what's going on there.