checking effects in baseline period
Following up on why there are significant correlations with dot_x starting from
the moment of dot presentation I explicitly checked whether the mean baseline
activity exhibits any task-related effects (that shouldn't be there). This is
done in the notebook baseline_correlations.ipynb. Additionally, I had a closer
look at the results of inference with single dot regressors specifically for the
very early period from dot onset. In short: The mean baseline signal correlates
with the x-coordinate of all 5 dots and most strongly with that of the first dot
while it doesn't, e.g., correlate significantly with the y-coordinate.
Correspondingly, I see a significant (but small) dot_x_1 effect throughout the
whole period from 0 to 500 ms after first dot onset when the baseline is not
subtracted. I do not believe that the MEG signal (brain) can predict which dot
will be coming up next, not even the sign of the dot x-coordinate, i.e.,
whether it will appear left or right of the middle of the screen. Hence, this
correlation must result from smoothing introduced by the filtering of the MEG
signal with a low-pass filter with cut-off at 45 Hz. For more discussion, see
below.

Single dot regressor results
----------------------------
Without baseline correction I still see the effects at 120 and 170 ms after dot
onset for each of the 5 dot regressors, but the peak is double as high at 120 ms
compared to 170 ms for dots 3 and 4. This is probably the reason why I also see
larger 120 ms peaks in the sequential inference. Most striking in these results
is, however, that the dot_x_1 betas appear to hover over their permuted
counterparts throughout the whole time period from 0 to 500 ms after dot onset,
i.e., absmax betas roughly stay above 0.04 while permuted absmax betas stay
around 0.02, as for all dot_x regressors.

How can permuted betas show very small, but consistent correlation with dot_x
regressors? This is most likely an artefact of taking the absolute of beta
values. There will always be some correlation and the mean across positive
random values will be a reasonably high positive number.

I checked the topography of the inferred betas for dot_x_1 at 0 ms. This is a
weak version of the typical centro-parietal pattern that I see for accumulated
evidence after 300 ms, indicating that the baseline effects mostly reflect the
sustained late signal, or maybe a kind of average signal across time points.

Absmax betas for the other dots also tend to be higher than their permuted
counterparts in the time period from 0 to 100 ms after dot onset, or even before
dot onset, but the effects are not as strong as for the first dot (difference
between betas and permuted betas is not as high).

Investigating the baseline period itself
----------------------------------------
In the new notebook baseline_correlations.ipynb I compute the mean MEG signal
within the period from -300 ms to 0 ms from first dot onset within each trial.
This is the measure that is then subtracted from the signal for baseline
correction. I then apply my standard single dot analysis to the mean baseline
signal to see whether it contains any task-relevant effects.

As my previous observations of differences between results with and without
baseline correction suggested, the baseline signal shows quite pronounced
effects for the dot_x regressors: the absmax betas are between 0.07 and 0.10
which is in the range of the effects 300 ms after dot onset for single dot
regressors. In comparison, the next largest effect in the baseline period is
that of entropy at 0.05 followed by response (0.04), dot_y_1 (0.04) and
abs_dot_x_1 (0.04). All other regressors had absmax betas of 0.03 or lower. It
is conceivable that response and entropy have effects before the onset of the
dots, because they may reflect the state of the individual at the beginning of
the trial, but, still, I don't see why there should be any correlation with
dot_x.

Discussion
----------
In the following I will discuss and try to exclude possible reasons for the
seen effects.

repeated trials
...............
Some trials were repeated, at least the first 4 dots of those trials in which we
manipulated the 5th dot x-coordinate. Couldn't it be that people actually
predicted which dot would occur? This is highly unlikely, because I see the
largest baseline period effect of dot_x for the first dot and the first dot
location should have never been predictable as the order of trials was
randomised.

no baseline effect of dot_y
...........................
If the filtering introduced these baseline effects why didn't it also introduce
baseline effects for dot_y? - dot_y has reasonably sized effects, but they are
not as strong as those for dot_x (maximum absmax values around 0.1 vs. 0.15). It
could simply be that the average across the trial was not large enough to induce
a filter-created image of dot_y in the baseline period. Also, dot_x may be
represented for much longer in the trial as it partially lives on in accumulated
evidence.

mechanism for filter distortions
................................
I currently don't understand how the filter can introduce this smearing of
correlations across the trial. The cut-off of 45 Hz of the low-pass filter has
a period of 22 ms which is relatively small compared to the 300 ms baseline
period + 100 ms offset until which a first peak of dot_x related activity
appears. Why would the filter distort the signal over so large time periods?

consequences
............
Assuming that the filter smeared out some kind of average of dot-related
effects across the trial what would be the consequences for my inferences?
Inspired by the effects seen for dot_x_1, I would guess that the distortion
just adds an offset to the beta values. By subtracting the baseline signal I get
rid of the offset as I observe with baseline correction. I wonder, though, how
far the smearing goes, i.e., if I subtract the baseline signal will it only
solve the problem for early dots, but leave some offset in late dots? On the one
hand, the offset in betas appears to be strongest for the first dot suggesting
that later dots have no severe issues. On the other hand, the first dot had
clearly the strongest effect in the baseline signal which may mean that the
baseline period does not contain all the smearing effects meaning that
subtracting the baseline signal would not completely remove the smearing and all
beta offsets.

Probably it would be best to try to get rid of the filter.