first single dot inference directly in source space
Because of the many regressors and time points single dot inference takes a
while. So I had it run over the weekend, but it broke with an error relating to
lost connection with ZIH after it had computed the non-permuted results. Thus, I
have so far not computed any permutations and will do that once I find the need.

Results are in:
	source_singledot_201706161257.h5

I did this mostly to check whether the dot-onset triggered effects persist in
the source space. - They do, i.e., for the regions strongly implicated in the
sequential inference (e.g., v23ab), I see the dot-specific effects clearly. This
is no surprise given the circularity of the argument, but it could still have
been that the sequential effects are driven by some strong effects for single
dots within the sequence. This appears not to be the case. One strange thing,
however, is that the first dot appears to have elevated effects that do not
necessarily show the same prototypical time course as the remaining dots.

Of the trial-wise regressors, entropy has a comparably high, sustained
signal somewhere or across the cortex. Trial_time seems to be represented
somewhere after 100 ms after first dot onset. Response, has only a weak average
correlation which rises slowly and noisily. Because of the missing permutations
I can't quite tell which of these effects are genuine. The areas which show
greatest correlations (or anti-correlations) with the trial regressors, however,
are quite telling. Entropy, for example, appears to correlate with signals in
dorsolateral prefrontal cortex, V6, an area in auditory association coretx
(STSda) and IP2 while the response has strongest effects in posterior
cingulate cortex, mid cingulate cortex and the dorsal visual stream. Trial_time,
i.e., steadily increasing or decreasing signals throughout the experiment, can
be found in a temporal-parietal-frontal network covering ventral stream visual
cortex (VMV2), the temporo-parieto-occipital junction (TPOJ), orbital and polar
frontal cortex (47m, 45), inferior frontal cortex (IFSp), insular cortex(FOP4)
and mid-cingulate motor areas (24dd, 24dv).