investigating new dot-level regressors
When reminded that dot_y also has significant effects, Stefan got worried that
dot_x does not only represent momentary evidence (I still don't understand how
he came to think otherwise). So what else does dot_x possibly stand for and what
for do we find correlates in the MEG signal? The brain could respond to a lot of
different stimulus features. I have now added a few:

- sum over x- or y-coordinate (sum_dot_[xy]) with a delayed version
  (sum_dot_[xy]_prev)
- distance of the shown dot to the centre of the screen (dot_dist)
- movement of the dot relative to its previous position separately for
  x and y-coordinates (move_[xy])
- distance of the shown dot relative to its previous position (move_dist)

My original idea was to add a subset of these regressors to a new extended
analysis in which I could better identify decision-related effects from pure
sensory dot-related effects. The new regressors, however, can be quite strongly
correlated with previous regressors. For example, sum_dot_x_prev is by
construction pretty much the same as accev (difference is only in the randomly
sampled values of bias that I add to accev). But also the move regressors are
strongly correlated with the original dot positions (typically r>0.6). This is
also true for the dot_dist. Only move_dist and sum_dot_y_prev are relatively
unrelated to themselves and previous regressors - so are fit for inclusion in a
new GLM. The remaining new regressors should be seen to contribute to effects
seen for the regressors that they correlate with.

I wonder how strong a correlation between two regressors can be before they
significantly distort their beta estimates? Maybe a correlation of 0.6 still
spans a wide enough space such that the effects of the involved regressors can
be identified?