Started sequential source GLM with new dot regressors
As per the discussion in the previous commit I added move_dist and
sum_dot_y_prev to the GLM and reran infer_sequential_source and
second_level_abs_inference. Computations are still running. As previously,
computing one permutation of GLM inference takes about 5 hours.

For the second_level_abs_inference I reset p_thresh to 0.02. I also renamed mu_t
to mu_z thinking that this describes the value better, but I realise that t- and
z- values are essentially the same and I should rather stay consistent with
previous results. Info: both t- and z-values are

	(data point - sample mean) / (sample std / sqrt(N))

where N is the number of data points in the sample. Technically for a z-value
the sample std is replaced with a known population standard deviation, but you
may still call the sample-based equation an approximation of the z-value. Proper
z-values are supposed to be Gaussian, while t-values are supposed to be
t-distributed under the null hypothesis.