new normalisation for regressors that are constant for 1st dot
As discussed in commit da83ddd8, I now treat normalisation of regressors special
that have a constant value for the 1st dot (these are the regressors which
depend on previous dots without depending on any trial-specifics such as the
response): I now do not use the constant 1st dot value in normalisation anymore
and set the 1st dot value to 0 after normalisation, effectively removing 1st dot
values from the analysis for these regressors.

Sidenote: I noticed that entropy and accev_cflip of the 5th dot are quite
strongly correlated (r~0.6 for many subjects). Consequently, I thought about
removing accev_cflip from analysis, but then decided against it, because across
the 5 dots the correlation is only around 0.3 (entropy is 0 for the first 4
dots here). Also accev_cflip is constrained by the first 4 dots while entropy is
not. So entropy should capture most of the common effect which is fine by me.