find_slabs_threshold can now exclude regressors
I noticed that the threshold was pretty high for one of the recent results and
wondered where it came from. It was due to the trial regressors in the
sequential analysis which tended to have relatively high baseline effects,
probably because they were only based on one dot instead of 4 or 5 as the dot
regressors. Because second_level_abs_inference uses directly the betas across
subjects without normalising by their variance, the increased variance of single
dot based trial regressors in comparison to 4 or 5 dot based dot regressors is
ignored on the second level and their increased betas also lead to larger effect
sizes on the second level, but only around the baseline. I should think about
taking perhaps t-values to the second level instead of directly the betas.

Anyway, find_slabs_threshold now allows to specify a set of regressors which can
be excluded from the computation of the threshold.