raw betas vs. beta-ts for second level: minimal differences?
In my last commit I wondered whether I should use the raw betas as before, or
the betas scaled by their standard error to avoid that trial regressors have
relatively high baseline effects (presumably driven by noise). So I investigated
how this would affect second level inferences by looking at the t-values that I
would compute on the second level and which is close to what I compute with the
folded normal measures. It turns out that these t-values, on average, will only
minimally differ when I either use betas or beta-t-values as basis for second
level inference. A few individual second level t-values may differ considerably,
though and specifically for entropy, which had the raised baseline on the second
level, differences are larger.

Clearly, the mean betas differ strongly from mean beta t-values, when averaging
across subjects, but this difference will also be reflected in the
corresponding standard deviation such that the second level t-values turn out to
be very similar. Given that the average second level t-values (across locations)
are above 1.0 for significant effects, the average difference between these
t-values for betas or beta-t-values as basis is quite small with values around
or smaller than 0.1 for different regressors. Only for entropy it's 0.17.

Also the distribution of the magnitude of these differences suggest that the
impact of normalising betas by their standard error has a small effect on
second level t-values: 90% of the difference magnitudes are below 0.2 for the
most important regressors and 99% of them are below 0.4. For entropy these
numbers are a bit higher: 0.38 and 0.58, respectively. In absolute numbers this
means that of 18100 differences typically less than 2000 have magnitudes larger
than 0.2 (for entropy: 6209).

These findings suggest that my GLM works reasonably well and that beta standard
errors do not differ strongly across regressors, time points and locations.

In conclusion, I do not expect large differences in results when using betas or
normalised betas as basis for second level inference, but my intuition still
says that I should rather use the normalised betas so that the information about
the uncertainty over betas does not get lost on the way to the second level.