investigating small variances across subjects + bugfix overlap
A plausible explanation for the large differences in variance across subjects
for areas which show an effect is that those areas with small variances
are particularly large areas which are only partially activated. If, for
example, an area consisting of 500 sources has an effect across all sources and
an area with 2000 sources has an equivalent effect in a subset of 500 of its
sources, the large area will show smaller variance and smaller effects when I
average across all the 2000 sources compared with the small area.

I checked this hypothesis by producing a scatter plot of average standard
deviations (across subjects) in areas vs. the number of vertices (sources) in
these areas. There is a clear anti-correlation between the two, as expected,
especially, large areas with more than 2000 vertices have mean standard
deviations below 0.02 while the majority of small areas has mean standard
deviations above 0.02 (some also have small standard deviations - as far as I
saw when there is consistently no effect across subjects).

This constitutes a problem: It makes the magnitude of mu (the mean effect) hard
to interpret. Put differently, why should I penalise large areas for showing
only a partial, but consistent and perhaps large activation? This is the problem
underlying the differences in results between mu_t and mu_p_large discussed in
the previous commit.

What are possible solutions? One possibility is to not average within an area,
but to take the maximum value from that area. This might be the better choice,
but I then worry that large areas will benefit by having a better chance of
randomly containing a large value due to noise. Alternatively, I can just
concentrate on the mu_t value which already showed me both kinds of effects.
However, because mu is only positive and the mu_t value cannot follow a
t-distribution, the mu_t value may be misinterpreted, or behave unintuitively.
Hence I thought of another measure which can quantify certainty in the
existence of an effect: mu_testval which is the value of mu for which 95% of all
posterior samples are larger than itself.

Properties of mu_testval: It, in principle, measures the existence of an effect
directly in units of mu. It should strongly correlate with mu_t, but is based
directly on the sampled distribution and should therefore be more reliable. It
adds to information provided by mu_p_large by concentrating on the low end of
the posterior instead of on the upper end. Currently, I'm pretty curious about
the distribution of mu_testval across regressors and whether the large areas
become 'significant' using it (first analysis still runs).

The introduction of mu_testval alleviates some of my issues and appears to be a
good solution for now, but if I want to look at effect magnitudes at some point
I should probably switch to taking the maximum across sources in an area instead
of the mean.

bugfix
------
overlap used the wrong standard deviation (that of the posterior of mu
instead of the mean sigma)! Previous values of overlap are therefore not
informative! Well, their interpretation is: If the posterior over mu was
Gaussian, what would be the probability that mu is negative? So this just
measures pretty much the same thing as mu_t. The original idea was that overlap
tells you something about the distribution of betas: The amount of overlap
between the two Gaussian components of the mixture distribution for the
posterior means of mu and sigma ranging between -inf (no overlap) and log(0.5)
(complete overlap).